+++ ui/widgets/account.h
@@ -24,6 +24,7 @@
 #include <QMap>
 #include <QString>
 #include <QVariant>
+#include <QStandardItemModel>
 
 #include "shared/global.h"
 
@@ -44,6 +45,9 @@ public:
     void setData(const QMap<QString, QVariant>& data);
     void lockId();
 
+private slots:
+    void onComboboxChange(int index);
+    
 private:
     QScopedPointer<Ui::Account> m_ui;
 };